fix(LottoService): use BigDecimal
use BigDecimal

+++ src/main/java/lotto/controller/LottoController.java
@@ -7,6 +7,8 @@ import lotto.dto.Result;
 import lotto.service.LottoService;
 import lotto.view.InputView;
 import lotto.view.OutputView;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 public class LottoController {
@@ -25,7 +27,7 @@ public class LottoController {
         Integer winningBonusNumber = getWinningBonusNumber();
 
         List<Integer> calculatedLottoResults = lottoService.calculateLottoResults(lottos, winningSixNumbers, winningBonusNumber);
-        Double earningRate = lottoService.calculateEarningRate(money.getMoney(), calculatedLottoResults);
+        BigDecimal earningRate = lottoService.calculateEarningRate(money.getMoney(), calculatedLottoResults);
         printCalculatedResult(new Result(calculatedLottoResults, earningRate));
 
         Console.close();
