+++ src/main/java/lotto/dto/Result.java
@@ -1,6 +1,7 @@
 package lotto.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
-public record Result(List<Integer> lottoMatchResults, Double earningRate) {
+public record Result(List<Integer> lottoMatchResults, BigDecimal earningRate) {
 }
